Moved util functions to util.go
